refactor and deletions - working
