+++ examples/menu.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# Example project menu, auto-detected by Wingman CLI if present in a project.
+# This example runs npm scripts and provides worktree helpers.
+
+echo "This is an example menu.sh. Customize per project."
+echo "Try: npm run start or npm run web"
+
