+++ package.json
@@ -10,12 +10,12 @@
     "web": "node scripts/start-with-scheduler.js server.js",
     "web:no-scheduler": "node server.js",
     "scheduler": "node scheduler/schedule.js",
-    "logs": "node log-observer.js",
-    "raw-logs": "node raw-log-viewer.js",
     "clearall": "node scripts/clearall.js",
     "worktree:switch": "./scripts/worktree-switch.sh",
     "worktree:create": "./scripts/worktree-create.sh",
-    "menu": "./menu.sh",
+    "menu": "bash examples/menu.sh",
+    "servers:init": "node scripts/init-builtin-servers.js",
+    "clean:artifacts": "node scripts/clean-artifacts.js",
     "db:migrate": "node scripts/migrate.js",
     "db:seed": "node scripts/seed.js",
     "test": "jest",
